+++ pyproject.toml
@@ -4,10 +4,9 @@ version = "0.1.0"
 description = "newapi.ai system check in"
 requires-python = ">=3.11"
 dependencies = [
-  "camoufox[geoip]>=0.4.11",
   "httpx[http2]>=0.24.0",
   "playwright>=1.40.0",
-  "python-dotenv>=1.0.0",
+  "python-dotenv>=1.0.0"
 ]
 
 [dependency-groups]
